+++ src/main/java/domain/Car.java
@@ -1,9 +1,13 @@
-package entity;
+package domain;
 
 public class Car {
 
     String name;
     int position = 0;
+
+    public Car(String name) {
+        this.name = name;
+    }
 
     public int getPosition() {
         return position;
@@ -13,5 +17,9 @@ public class Car {
         if (random >= 4) {
             this.position += 1;
         }
+    }
+
+    public Boolean isWinner(int winningPosition) {
+        return this.position == winningPosition;
     }
 }
